requirements.txt
discord.py
flask
PyNaCl
yt-dlp